+++ CHANGES.md
@@ -1,9 +1,11 @@
 # Updates
 
-## Upcoming
+## 1.0.3 (2021-11-19)
 
 - Add sphinx documentation for Read the Docs
 - Fix pyproject.toml to match [PEP 621](https://www.python.org/dev/peps/pep-0621/)
+- Add property dwd.direction
+- Pin dependency versions to support `pip install --require-hashes`
 
 ## 1.0.2 (2021-05-28)